Cambios con respecto al calculo total de objetivos de entero a double

+++ Bizapps-evaluacion/src/interfacedao/transacciones/IUtilidadDAO.java
@@ -4,8 +4,8 @@ public interface IUtilidadDAO {
 	
 	public void eliminarConductaPorCompetencia(Integer eva, Integer com);
 	public void eliminarConductaPorEvaluacion(Integer eva);
-	public String obtenerValoracionFinal(Integer resultado);
-	public String obtenerValoracionFinalSimple(Integer resultado);
+	public String obtenerValoracionFinal(double resultado);
+	public String obtenerValoracionFinalSimple(double resultado);
 	public void eliminarCompetenciaPorEvaluacion(Integer eva);
 	public void eliminarCapacitacionPorEvaluacion(Integer eva);
